+++ saleor/plugins/tests/test_manager.py
@@ -6,9 +6,11 @@ from django.http import HttpResponseNotFound, JsonResponse
 from django_countries.fields import Country
 from prices import Money, TaxedMoney
 
+from ...checkout import CheckoutLineInfo
 from ...checkout.utils import fetch_checkout_lines
 from ...core.taxes import TaxType
 from ...payment.interface import PaymentGateway
+from ...product.models import Product
 from ..manager import PluginsManager, get_plugins_manager
 from ..models import PluginConfiguration
 from ..tests.sample_plugins import (
@@ -121,6 +123,92 @@ def test_manager_calculates_checkout_line_total(
         [discount_info],
     )
     assert TaxedMoney(expected_total, expected_total) == taxed_total
+
+
+def test_manager_get_checkout_line_tax_rate_sample_plugin(
+    checkout_with_item, discount_info
+):
+    line = checkout_with_item.lines.all()[0]
+    plugins = ["saleor.plugins.tests.sample_plugins.PluginSample"]
+    unit_price = TaxedMoney(Money(12, "USD"), Money(15, "USD"))
+
+    variant = line.variant
+    checkout_line_info = CheckoutLineInfo(
+        line=line,
+        variant=variant,
+        channel_listing=variant.channel_listings.first(),
+        product=variant.product,
+        collections=[],
+    )
+
+    tax_rate = PluginsManager(plugins=plugins).get_checkout_line_tax_rate(
+        checkout_with_item,
+        checkout_line_info,
+        checkout_with_item.shipping_address,
+        [discount_info],
+        unit_price,
+    )
+    assert tax_rate == Decimal("0.08")
+
+
+@pytest.mark.parametrize(
+    "unit_price, expected_tax_rate",
+    [
+        (TaxedMoney(Money(12, "USD"), Money(15, "USD")), Decimal("0.25")),
+        (Decimal("0.0"), Decimal("0.0")),
+    ],
+)
+def test_manager_get_checkout_line_tax_rate_no_plugins(
+    checkout_with_item, discount_info, unit_price, expected_tax_rate
+):
+    line = checkout_with_item.lines.all()[0]
+    variant = line.variant
+    checkout_line_info = CheckoutLineInfo(
+        line=line,
+        variant=variant,
+        channel_listing=variant.channel_listings.first(),
+        product=variant.product,
+        collections=[],
+    )
+    tax_rate = PluginsManager(plugins=[]).get_checkout_line_tax_rate(
+        checkout_with_item,
+        checkout_line_info,
+        checkout_with_item.shipping_address,
+        [discount_info],
+        unit_price,
+    )
+    assert tax_rate == expected_tax_rate
+
+
+def test_manager_get_order_line_tax_rate_sample_plugin(order_with_lines):
+    order = order_with_lines
+    line = order.lines.first()
+    product = Product.objects.get(name=line.product_name)
+    plugins = ["saleor.plugins.tests.sample_plugins.PluginSample"]
+    unit_price = TaxedMoney(Money(12, "USD"), Money(15, "USD"))
+    tax_rate = PluginsManager(plugins=plugins).get_order_line_tax_rate(
+        order, product, None, unit_price,
+    )
+    assert tax_rate == Decimal("0.08")
+
+
+@pytest.mark.parametrize(
+    "unit_price, expected_tax_rate",
+    [
+        (TaxedMoney(Money(12, "USD"), Money(15, "USD")), Decimal("0.25")),
+        (Decimal("0.0"), Decimal("0.0")),
+    ],
+)
+def test_manager_get_order_line_tax_rate_no_plugins(
+    order_with_lines, unit_price, expected_tax_rate
+):
+    order = order_with_lines
+    line = order.lines.first()
+    product = Product.objects.get(name=line.product_name)
+    tax_rate = PluginsManager(plugins=[]).get_order_line_tax_rate(
+        order, product, None, unit_price,
+    )
+    assert tax_rate == expected_tax_rate
 
 
 @pytest.mark.parametrize(